requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
selenium>=4.0.0
pydantic>=1.8.0
python-dotenv>=0.19.0
celery>=5.3.0
redis>=4.5.0
fastapi-cache2[redis]>=0.2.1 